Modifications to see the stdout of ioIface
ioIface program was launched with Popen from python.
The stdout of ioI face was redirected to a PIPE and the pipe
wasn't be read by anybody.
I suppose that it was the problem that was happening when the
controller get freeze after working for long time. (The PIPE
became very big and nobody read it).
Now the stdout of ioIface is redirected to a file.
To do this the ioIface C program was be modified removing the stdout
and stderr buffer. (By default the standard C library buffers the
stdout and stderr when not sending them to a terminal)