Delete events when deleting a person or a door involved in the event
In some situations, when a visitor exits through a visit exit door,
the controller receives the order of deleting the person and it
doesn't still delete the event previously sent. In this situation
the controller wasn't able to delete the visitor.
In this fix, in the declaration of "Event" table, ON CASCADE DELETE
was added to personId and also to doorId to be coherent.
This situation was seen when the there wasn't communication between
controller and server an a lot of visit exit event were accumulated