Deleted persons are renamed as name + (DELETED)
When a deleted person is committed (all the controllers
confirms the deletion), the "resStateId" of the person
is assigned the DELETED value, the name of the person is
ranamed as NAME + (DELETED), "cardNumber" and "identNumber"
are set to NULL to permit reuse of the card and create
again the person with the same "identNumber".
For this change, now "cardNumber" and "identNumber" can
be NULL in db schema.