Show person on events of person passing through not allowed door
The events coming from the controller have the card number instead of
the person id because the person trying to pass a door controlled by this controller,
couldn't be in this controller (no access in any of the doors controlled by it) but
the person could be in the central data base. In this situation it is desirable to
show the involved person in the event, also when they are denyied.
cardNum2PrsnId() function is used to put the person id in the event dictionary and
remove the card number.

-saveEvent() method and Event table was modified to save card number instead of
 personId when no connection to server.

-canAccess() method was modified not returning the person id.

-Fix bug of not deleting person with id = 1 (This was from the time that id = 1
 was reserved for visits or something like that)

-Delete some useless lines commented